requirements.txt
environ==1.0
django-environ==0.4.5
Django==2.1.3
django-heroku==0.3.1
djangorestframework==3.9.0
django-url-filter==0.3.10
django-debug-toolbar==1.10.1
gunicorn==19.9.0
dj-database-url==0.5.0
psycopg2-binary==2.7.6.1
Pillow==5.3.0